Payment.Providerをコンポーネントに切り出し
