+++ frontend/src/components/pages/Home.tsx
@@ -1,4 +1,4 @@
-import { memo, useContext, useEffect, VFC } from 'react'
+import { memo, useContext, VFC } from 'react'
 import { Link } from 'react-router-dom'
 
 import { Box, Center } from '@chakra-ui/react'
@@ -13,70 +13,10 @@ import { PaymentListArea } from 'components/organisms/PaymentListArea'
 import { HomeHeaderLayout } from 'components/templates/HomeHeaderLayout'
 import { AuthContext } from 'context/AuthContext'
 import { PaymentContext } from 'context/PaymentContext'
-import { getPayments } from 'lib/api/payment'
-import { getTeamStatus } from 'lib/api/team'
-import { auth } from 'lib/firebase'
-import { useToast } from 'lib/toast'
 
 export const Home: VFC = memo(() => {
-  const {
-    paymentList,
-    setPaymentList,
-    isPaymentListLoaded,
-    setIsPaymentListLoaded,
-    teamStatus,
-    setTeamStatus,
-    isTeamStatusLoaded,
-    setIsTeamStatusLoaded,
-  } = useContext(PaymentContext)
+  const { isPaymentListLoaded, teamStatus, isTeamStatusLoaded } = useContext(PaymentContext)
   const { currentUser } = useContext(AuthContext)
-  const { errorToast } = useToast()
-
-  const handleGetPayments = async () => {
-    const idToken = await auth.currentUser?.getIdToken(true)
-    try {
-      const res = await getPayments(idToken)
-      if (res?.status === 200) {
-        setPaymentList(res?.data)
-      } else {
-        errorToast('取得に失敗しました')
-      }
-    } catch {
-      errorToast('取得に失敗しました')
-    }
-    setIsPaymentListLoaded(true)
-  }
-
-  const handleGetTeamStatus = async () => {
-    const idToken = await auth.currentUser?.getIdToken(true)
-    if (currentUser) {
-      try {
-        const res = await getTeamStatus(currentUser.teamId, idToken)
-        if (res?.status === 200) {
-          setTeamStatus(res?.data)
-        } else {
-          errorToast('取得に失敗しました')
-        }
-      } catch {
-        errorToast('取得に失敗しました')
-      }
-      setIsTeamStatusLoaded(true)
-    }
-  }
-
-  useEffect(() => {
-    handleGetPayments().catch((err) => {
-      console.log(err)
-    })
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
-  useEffect(() => {
-    handleGetTeamStatus().catch((err) => {
-      console.log(err)
-    })
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [paymentList])
 
   return (
     <>
